requirements.txt
python-kasa
python-dotenv
pytz
zmanim
fastapi
uvicorn